split logging into separate class
